chore: move seelog custom formatter to log package
